[Chore] 모임 삭제시 매니저만 삭제권한 주도록 분기처리

+++ back/src/routes/club/router.js
@@ -270,8 +270,9 @@ clubRouter.put("/:id", verifyToken, async (req, res) => {
 // 없는 모임을 삭제할 경우, 에러 처리
 clubRouter.delete("/:id", verifyToken, async (req, res, next) => {
   try {
+    const user_id = req.user;
     const club_id = req.params.id;
-    const deletedClub = await clubService.deleteClub({ club_id });
+    const deletedClub = await clubService.deleteClub({ club_id, user_id });
 
     if (deletedClub.errorMessage) {
       res.status(403).json({ success: false, err: deletedClub.errorMessage });
